requirements.txt
# Google Cloud Platform libraries
google-cloud-aiplatform
google-cloud-documentai
google-cloud-storage
google-cloud-bigquery

# Imaging and PDF manipulation
Pillow  # PIL Fork
PyPDF2

# Data visualization libraries
seaborn
matplotlib

# Machine Learning library
scikit-learn  # sklearn
